appointment-service/requirements.txt
Django
djangorestframework
psycopg2-binary
pika
requests
PyJWT